Move `DataGridViewChooseDataSourceActionList` into a separate file
Related #13231

## Root Cause

- Missing

## Proposed changes

- Move `DataGridViewChooseDataSourceActionList` into a separate file
- Add comment about missing structure for choosing data source designer
action item.

## Customer Impact

- None

## Regression?

- No

## Risk

- Minimal

## Test environment(s)

- 10.0.100-preview.3.25201.16
